Поменял провайдер в NgModule, теперь отправляется объект у которого точка входа DataService, но можно передавать разную реализацию
